move listeners into Links.js
